rename Mobile Panel to RichData
