+++ src/components/HURUmap/Panel/MobilePanel/RichData.js
